Added functions.php
