+++ index.php
@@ -1,35 +1,6 @@
 <?php
     require_once('helpers.php');
-
-    /**
-     * Урезает оригинальный текст, если его длина меньше заданного числа символов
-     *
-     * @param string $content Оригинальный текст
-     * @param int $lengthLimit Лимит на число символов
-     *
-     * @return string Возвращается либо урезанный текст с ссылкой, либо оригинальный
-     */
-    function trimContent(string $content, int $lengthLimit = 300): string
-    {
-        $moreLink = '<a class="post-text__more-link" href="#">Читать далее</a>';
-        $contentWords = explode(' ', $content);
-        $contentLength = 0;
-        $trimmedWords = [];
-
-        foreach ($contentWords as $word) {
-            $contentLength += mb_strlen($word) + 1;
-
-            if ($contentLength <= $lengthLimit) {
-                $trimmedWords[] = $word;
-            } else {
-                break;
-            }
-        }
-
-        $trimmedContent = implode(' ', $trimmedWords);
-
-        return $contentLength > $lengthLimit ? "<p>{$trimmedContent}...</p>{$moreLink}" : "<p>{$content}</p>";
-    }
+    require_once('functions.php');
 
     $pageContent = include_template('main.php', [
         'popularPosts' => [
